* make protocol and supertoken page network specific * remember selected network when switching page * redirect previous URLs to matic * make token and symbol search case insensitive TODO: adjust tests
